requirements.txt
aiohttp==3.7.4.post0
jsonslicer==0.1.6
simplejson==3.17.2
python-dateutil==2.8.1
